Create border-control.php
Removed color option from this group control. Because color can be groped inside other color related settings of a widget. Otherwise, one color would be inside border control and if you wanted to change the color for hover state, then you would have to put that color inside other section. this is bad experience. So, I removed color from this control so that color can be displayed in an appropriate place.